move main scraper parser to a separate script
