+++ scraper/utilities/parsers/__init__.py
@@ -48,208 +48,6 @@ from pathlib import Path
 from typing import List, Optional
 
 
-def build_parser(argv: Optional[List] = None):
-    """Return ArgumentParser parser for script."""
-    # Create parser object
-    parser = argparse.ArgumentParser(
-        prog="cazy_webscraper.py",
-        description="Scrapes the CAZy database",
-        formatter_class=argparse.ArgumentDefaultsHelpFormatter,
-    )
-
-    # Add optional arguments to parser
-
-    # Add option to specify path to configuration file
-    parser.add_argument(
-        "-c",
-        "--config",
-        type=Path,
-        metavar="config file",
-        default=None,
-        help="Path to configuration file. Default: None, scrapes entire database",
-    )
-
-    # Add option to define complete classes to scrape
-    parser.add_argument(
-        "--classes",
-        type=str,
-        default=None,
-        help="Classes from which all families are to be scraped. Separate classes by ','"
-    )
-
-    # Add option to use own CAZy class synoymn dict
-    parser.add_argument(
-        "--cazy_synonyms",
-        type=Path,
-        default=None,
-        help="Path to JSON file containing CAZy class synoymn names",
-    )
-
-    # Add option to display citation
-    parser.add_argument(
-        "-C",
-        "--citation",
-        dest="citation",
-        action="store_true",
-        default=False,
-        help="Print cazy_webscraper citation message",
-    )  
-
-    # Add option to provide a path to an existing SQL database to add newly scraped data to
-    parser.add_argument(
-        "-d",
-        "--db_output",
-        type=Path,
-        default=None,
-        help="Target output path to build new SQL database",
-    )
-
-    parser.add_argument(
-        "-D",
-        "--database",
-        type=Path,
-        default=None,
-        help="Path to an existing local CAZy SQL database",
-    )
-
-    parser.add_argument(
-        "--dict_output",
-        type=Path,
-        default=None,
-        help="Target output path to write out JSON file of protein accesisons and CAZy family annotations",
-    )
-
-    # Add option to force file over writting
-    parser.add_argument(
-        "-f",
-        "--force",
-        dest="force",
-        action="store_true",
-        default=False,
-        help="Force file over writting",
-    )
-
-    # Add option to specify families to scrape
-    parser.add_argument(
-        "--families",
-        type=str,
-        default=None,
-        help="Families to scrape. Separate families by commas 'GH1,GH2' (case sensitive)"
-    )
-
-    # Add option to restrict scrape to specific genera
-    parser.add_argument(
-        "--genera",
-        type=str,
-        default=None,
-        help="Genera to restrict the scrape to"
-    )
-
-    # Add log file name option
-    # If not given, no log file will be written out
-    parser.add_argument(
-        "-l",
-        "--log",
-        type=Path,
-        metavar="log file name",
-        default=None,
-        help="Defines log file name and/or path",
-    )
-
-    # Add option to enable retrieval of subfamilies
-    parser.add_argument(
-        "--no_db",
-        dest="no_db",
-        action="store_true",
-        default=False,
-        help="When called, disables building an SQL database",
-    )
-
-    # Add option to not delete content in the existing cache dir
-    parser.add_argument(
-        "--nodelete",
-        dest="nodelete",
-        action="store_true",
-        default=False,
-        help="When called, content in the existing cache dir is NOT deleted",
-    )
-
-    # Add option to enable number of times to retry scraping
-    parser.add_argument(
-        "-r",
-        "--retries",
-        type=int,
-        default=10,
-        help="Number of times to retry scraping a family or class page if error encountered",
-    )
-
-    # Add option to enable retrieval of subfamilies
-    parser.add_argument(
-        "-s",
-        "--subfamilies",
-        dest="subfamilies",
-        action="store_true",
-        default=False,
-        help="Enable retrieval of subfamilies from CAZy",
-    )
-
-    # Add option to restrict the scrape to specific species. This will scrape CAZymes from
-    # all strains belonging to each listed species
-    parser.add_argument(
-        "--species",
-        type=str,
-        default=None,
-        help="Species (written as Genus Species) to restrict the scrape to"
-    )
-
-    # Add option to restrict scraping to specific strains of organisms
-    parser.add_argument(
-        "--strains",
-        type=str,
-        default=None,
-        help=(
-            "Specific strains of organisms to restrict the scrape to "
-            "(written as Genus Species Strain)"
-        ),
-    )
-
-    # Add option to define time out limit for trying to connect to CAZy
-    parser.add_argument(
-        "-t",
-        "--timeout",
-        type=int,
-        default=45,
-        help="Connection timeout limit (seconds)"
-    )
-
-    # Add option for more detail (verbose) logging
-    parser.add_argument(
-        "-v",
-        "--verbose",
-        dest="verbose",
-        action="store_true",
-        default=False,
-        help="Set logger level to 'INFO'",
-    )
-    
-    # Add option to display version
-    parser.add_argument(
-        "-V",
-        "--version",
-        dest="version",
-        action="store_true",
-        default=False,
-        help="Print cazy_webscraper version number",
-    )    
-
-    if argv is None:
-        # parse command-line
-        return parser
-    else:
-        # return namespace
-        return parser.parse_args(argv)
-
-
 def build_genbank_sequences_parser(argv: Optional[List] = None):
     """Return ArgumentParser parser for the script 'expand.genbank_sequences.py'."""
     # Create parser object
